@@ -2,9 +2,14 @@ package org.isobit.app;
 
 import java.util.HashMap;
 
-import javax.ws.rs.core.Response;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.ext.ExceptionMapper;
+import jakarta.ws.rs.ext.Provider;
+
+/*import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
+*/
 
 @Provider
 public class AppExceptionMapper implements ExceptionMapper<Throwable> {
