@@ -3,43 +3,28 @@ package gob.regionancash.bpm;
 import java.util.List;
 
 import org.hibernate.QueryException;
-import org.hibernate.dialect.MySQL5Dialect;
-import org.hibernate.dialect.function.SQLFunction;
+import org.hibernate.dialect.MySQL8Dialect;
+//import org.hibernate.dialect.MySQL5Dialect;
+//import org.hibernate.dialect.function.SQLFunction;
 import org.hibernate.dialect.function.StandardSQLFunction;
 import org.hibernate.engine.spi.Mapping;
 import org.hibernate.engine.spi.SessionFactoryImplementor;
 import org.hibernate.type.StandardBasicTypes;
-import org.hibernate.type.StringType;
+//import org.hibernate.type.StringType;
 import org.hibernate.type.Type;
 
 
-public class MySQLCustomDialect extends MySQL5Dialect {
-	
+public class MySQLCustomDialect extends MySQL8Dialect {
+	/* 
 	class ListAggFunction implements SQLFunction {
 
-	    /**
-	     * The pattern that describes how the function is build in SQL.
-	     *
-	     * Replacements:
-	     * {path} - is replaced with the path of the list attribute
-	     * {separator} - is replaced with the separator (defaults to '')
-	     * {orderByPath} - is replaced by the path that is used for ordering the elements of the list
-	     */
 	    private String pattern;
 
-	    /**
-	     * Creates a new ListAggFunction definition which uses the ANSI SQL:2016 syntax.
-	     */
 	    public ListAggFunction() {
 	        this("GROUP_CONCAT({path} ORDER BY {orderByPath} SEPARATOR {separator})");
 	        //this("GROUP_CONCAT(DISTINCT {path} ORDER BY {orderByPath} SEPARATOR {separator})");
 	    }
 
-	    /**
-	     * Creates a new ListAggFunction definition which uses a database specific syntax.
-	     *
-	     * @param pattern  The pattern that describes how the function is build in SQL.
-	     */
 	    public ListAggFunction(String pattern) {
 	        this.pattern = pattern;
 	    }
@@ -86,5 +71,5 @@ public class MySQLCustomDialect extends MySQL5Dialect {
             );
         this.registerKeyword("SEPARATOR");
     }
-    
+    */
 }
