@@ -1,4 +1,4 @@
-package gob.regionancash.bpm;
+package gob.regionancash;
 
 import io.quarkus.test.junit.QuarkusTest;
 import org.junit.jupiter.api.Test;
@@ -6,13 +6,13 @@ import org.junit.jupiter.api.Test;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.is;
 
-//@QuarkusTest
-public class ResourceTest {
+@QuarkusTest
+public class GreetingResourceTest {
 
-    //@Test
+    @Test
     public void testHelloEndpoint() {
         /*given()
-          .when().get("/")
+          .when().get("/hello")
           .then()
              .statusCode(200)
              .body(is("Hello from RESTEasy Reactive"));*/
